API and html Raw DoTest and TestResult of Student

+++ Backend/Models/UserTest.cs
@@ -8,9 +8,9 @@ namespace Backend.Models
         public int UserTestId { get; set; }
         public int? TestId { get; set; }
         public int? UserId { get; set; }
-        public int? Point { get; set; }
+        public double? Point { get; set; }
         public DateTime? TestDate { get; set; }
-        public int? TotalTime { get; set; }
+        public double? TotalTime { get; set; }
 
         public virtual Test? Test { get; set; }
         public virtual User? User { get; set; }
